refactor: imports, perf-logging
