+++ bin/simulation.rs
@@ -6,13 +6,13 @@
 #![allow(unused_parens)]
 #![allow(private_bounds)]
 
-use myco_rs::error::MycoError;
-use myco_rs::network::{LocalServer1Access, LocalServer2Access};
 use myco_rs::{
-    calculate_bucket_usage,
     client::Client,
     constants::{DB_SIZE, DELTA, NUM_CLIENTS},
+    crypto::calculate_bucket_usage,
     dtypes::Key,
+    error::MycoError,
+    network::{LocalServer1Access, LocalServer2Access},
     server1::Server1,
     server2::Server2,
 };
